added InputField to components/ui so that it can be used throughout the project
